♻️ Move author summary view to upload route, revert shared AuthorField
